Added Initial Crawling Function
Current the crawler goes to the given seed links, gets the links from the webpages, and crawls to the new links.
It remembers the links it has been to and does not add those links to the frontier again
It only travels to links that are in the correct domains as given in the project description.
It removes queries and fragments before storing and traversing to new links
It does not store any further information besides the links it has seen as of right now